Finish QLDaoTao, init QLGiangVien, QLHocVien

+++ EnrolledManagementSystem/DTO/Requests/LopHocCreate.cs
@@ -15,7 +15,7 @@ namespace EnrolledManagementSystem.DTO.Requests
         [Required]
         public int SoLuongHocVien { get; set; }
         [Required]
-        public double HocPhi { get; set; }
+        public decimal HocPhi { get; set; }
         public string? MoTa { get; set; }
         public string? HinhAnh { get; set; }
     }
